Refactor recordings and add data dir for SVC Pro compatibility
