+++ backend_py/src/models/recordings.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+
+class RecordingInfo(BaseModel):
+    path: str
+    name: str
+    format: str
+    duration: str
+    size: str
+    created: str
